fix: 32-bit CI - AppendMenuA uIDNewItem must be uintptr_t not unsigned long long
The stdcall suffix was _AppendMenuA@20 (8-byte param) but user32.lib
exports _AppendMenuA@16 (4-byte UINT_PTR on x86). Also use GetTickCount64
unconditionally (was undeclared GetTickCount on 32-bit). 32-bit 14/14
tests now pass locally; unblocks CI which failed for 25+ consecutive runs.

+++ pbcompiler/runtime/pb_runtime.c
@@ -63,7 +63,7 @@ __declspec(dllimport) void* __stdcall GetStdHandle(unsigned int nStdHandle);
 __declspec(dllimport) int __stdcall SetConsoleTextAttribute(void* hConsoleOutput, unsigned short wAttributes);
 __declspec(dllimport) void* __stdcall CreateMenu(void);
 __declspec(dllimport) void* __stdcall CreatePopupMenu(void);
-__declspec(dllimport) int __stdcall AppendMenuA(void* hMenu, unsigned int uFlags, unsigned long long uIDNewItem, const char* lpNewItem);
+__declspec(dllimport) int __stdcall AppendMenuA(void* hMenu, unsigned int uFlags, uintptr_t uIDNewItem, const char* lpNewItem);
 __declspec(dllimport) int __stdcall DeleteMenu(void* hMenu, unsigned int uPosition, unsigned int uFlags);
 __declspec(dllimport) int __stdcall DestroyMenu(void* hMenu);
 __declspec(dllimport) void* __stdcall CreateCompatibleDC(void* hdc);
